partial steps towards upgrade tower popup
changed GunT1 to StaticBody2d and that introduced errors having to do with how Guns are named when added to the scene.  They are named GunT1, GunT2, GunT3, GunT4 etc.  There is also an issue where the Range collision shape gets triggered by my mouse (the _on_body_entered signal).  This is addressed by checking the name of the body that entered and if it is "DragTower" or begins with "Gun", ignore it and Return.